refactor: Rename `test_helpers.rs` to `helpers.rs`
