+++ tests/integration_test.rs
@@ -1,2 +1,2 @@
 mod commands;
-mod test_helpers;
+mod helpers;
